refactor : extract common method of token provider
